Sessions work yay. Finishing up pre-game stuff
Sessions have been fixed finally. Code has been reworked for a much
easier to read pre-game session. It's not complete, but it will do for
now to get into working on game logic. Timer element has been added for
the room and as a DIV in HTML. It's not complete nor even tested yet.